Changes for arm64 and Clang redefined-types

+++ Libraries/ProtocolBuffers/TextFormat.m
@@ -86,7 +86,7 @@ BOOL isHex(unichar c) {
     if (isSigned) {
       result = strtol(in_string, &out_string, 0);
     } else {
-      result = convertUInt32ToInt32(strtoul(in_string, &out_string, 0));
+      result = convertUInt32ToInt32((unsigned int)strtoul(in_string, &out_string, 0));
     }
   }
